refactor: make_habitat_configs.py -> simulators/habitat/configs.py
