+++ benchmarks/configs/monty_world_experiments.py
@@ -35,14 +35,14 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     get_env_dataloader_per_object_by_idx,
     get_object_names_by_idx,
 )
-from tbp.monty.frameworks.config_utils.make_habitat_configs import (
-    PatchViewFinderMontyWorldMountHabitatDatasetArgs,
-)
 from tbp.monty.frameworks.environments import embodied_data as ED
 from tbp.monty.frameworks.environments.two_d_data import NUMENTA_OBJECTS
 from tbp.monty.frameworks.experiments import MontyObjectRecognitionExperiment
 from tbp.monty.frameworks.models.sensor_modules import (
     DetailedLoggingSM,
+)
+from tbp.monty.simulators.habitat.configs import (
+    PatchViewFinderMontyWorldMountHabitatDatasetArgs,
 )
 
 from .ycb_experiments import (
