+++ benchmarks/configs/ycb_experiments.py
@@ -35,13 +35,6 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     get_env_dataloader_per_object_by_idx,
     get_object_names_by_idx,
 )
-from tbp.monty.frameworks.config_utils.make_habitat_configs import (
-    FiveLMMountHabitatDatasetArgs,
-    NoisySurfaceViewFinderMountHabitatDatasetArgs,
-    PatchViewFinderMountHabitatDatasetArgs,
-    PatchViewFinderMultiObjectMountHabitatDatasetArgs,
-    SurfaceViewFinderMountHabitatDatasetArgs,
-)
 from tbp.monty.frameworks.environments import embodied_data as ED
 from tbp.monty.frameworks.environments.ycb import (
     DISTINCT_OBJECTS,
@@ -58,6 +51,13 @@ from tbp.monty.frameworks.models.goal_state_generation import (
 from tbp.monty.frameworks.models.sensor_modules import (
     DetailedLoggingSM,
     FeatureChangeSM,
+)
+from tbp.monty.simulators.habitat.configs import (
+    FiveLMMountHabitatDatasetArgs,
+    NoisySurfaceViewFinderMountHabitatDatasetArgs,
+    PatchViewFinderMountHabitatDatasetArgs,
+    PatchViewFinderMultiObjectMountHabitatDatasetArgs,
+    SurfaceViewFinderMountHabitatDatasetArgs,
 )
 
 """
