+++ tests/unit/evidence_lm_test.py
@@ -39,15 +39,6 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     ExperimentArgs,
     PredefinedObjectInitializer,
 )
-from tbp.monty.frameworks.config_utils.make_habitat_configs import (
-    EnvInitArgsFiveLMMount,
-    EnvInitArgsPatchViewMount,
-    EnvInitArgsTwoLMDistantStackedMount,
-    FiveLMMountHabitatDatasetArgs,
-    NoisyPatchViewFinderMountHabitatDatasetArgs,
-    PatchViewFinderMountHabitatDatasetArgs,
-    TwoLMStackedDistantMountHabitatDatasetArgs,
-)
 from tbp.monty.frameworks.config_utils.policy_setup_utils import (
     make_informed_policy_config,
 )
@@ -66,6 +57,15 @@ from tbp.monty.frameworks.models.sensor_modules import (
     HabitatDistantPatchSM,
 )
 from tbp.monty.frameworks.utils.logging_utils import load_models_from_dir
+from tbp.monty.simulators.habitat.configs import (
+    EnvInitArgsFiveLMMount,
+    EnvInitArgsPatchViewMount,
+    EnvInitArgsTwoLMDistantStackedMount,
+    FiveLMMountHabitatDatasetArgs,
+    NoisyPatchViewFinderMountHabitatDatasetArgs,
+    PatchViewFinderMountHabitatDatasetArgs,
+    TwoLMStackedDistantMountHabitatDatasetArgs,
+)
 from tests.unit.resources.unit_test_utils import BaseGraphTestCases
