+++ tests/unit/graph_learning_test.py
@@ -37,14 +37,6 @@ from tbp.monty.frameworks.config_utils.make_dataset_configs import (
     ExperimentArgs,
     PredefinedObjectInitializer,
 )
-from tbp.monty.frameworks.config_utils.make_habitat_configs import (
-    EnvInitArgsFiveLMMount,
-    EnvInitArgsPatchViewMount,
-    EnvInitArgsSurfaceViewMount,
-    FiveLMMountHabitatDatasetArgs,
-    PatchViewFinderMountHabitatDatasetArgs,
-    SurfaceViewFinderMountHabitatDatasetArgs,
-)
 from tbp.monty.frameworks.config_utils.policy_setup_utils import (
     make_informed_policy_config,
 )
@@ -60,6 +52,14 @@ from tbp.monty.frameworks.utils.follow_up_configs import (
 from tbp.monty.frameworks.utils.logging_utils import (
     deserialize_json_chunks,
     load_stats,
+)
+from tbp.monty.simulators.habitat.configs import (
+    EnvInitArgsFiveLMMount,
+    EnvInitArgsPatchViewMount,
+    EnvInitArgsSurfaceViewMount,
+    FiveLMMountHabitatDatasetArgs,
+    PatchViewFinderMountHabitatDatasetArgs,
+    SurfaceViewFinderMountHabitatDatasetArgs,
 )
 from tests.unit.resources.unit_test_utils import BaseGraphTestCases
